app-server/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
pyOpenSSL==22.0.0
requests==2.27.1
supervisor==4.2.4
